split up siren classes into seperate files
