+++ .eslintrc.js
@@ -32,6 +32,7 @@ module.exports = {
     'global-require': 0,
     'guard-for-in': 0,
     'import/extensions': 0,
+    'import/prefer-default-export': 0,
     'indent': ['warn', 2],
     'keyword-spacing': 'error',
     'max-classes-per-file': 0,
@@ -41,8 +42,8 @@ module.exports = {
     'no-confusing-arrow': 0,
     'no-continue': 0,
     'no-extra-semi': 'error',
-    'no-mixed-requires': 0,
     'no-mixed-operators': 0,
+    'no-mixed-requires': 0,
     'no-multi-assign': 0,
     'no-param-reassign': 0,
     'no-plusplus': 0,
@@ -60,8 +61,8 @@ module.exports = {
     'object-curly-newline': 0,
     'prefer-destructuring': 1,
     'prefer-promise-reject-errors': 1,
-    'quotes': [2, 'single'],
     'quote-props': 0,
+    'quotes': [2, 'single'],
     'require-jsdoc': 0,
     'semi': ['error', 'always'],
     'space-before-blocks': 'error',
