+++ demo/box.htm
@@ -8,9 +8,9 @@
 <body>
   <div id="ship-output"></div>
   <script type="module">
-    import box  from './box.js';
-    import { Shipwreck } from '../src/shipwreck.js';
-    import { SirenEntity } from '../src/siren.js';
+    import box  from './box';
+    import { Shipwreck } from '../src/shipwreck';
+    import { SirenEntity } from '../src/lib/siren/index.js';
     const shipOutput = document.getElementById('ship-output');
     const ship = new Shipwreck(shipOutput);
     ship.entity = new SirenEntity(box);
