+++ src/lib/siren/index.js
@@ -1,0 +1,14 @@
+/**
+ * Classes for working with Siren Entities
+ *
+ * Entity
+ * - SubEntity
+ * - Links
+ * - Actions
+ *   - Fields
+ */
+
+export { SirenAction } from './siren-action.js';
+export { SirenEntity } from './siren-entity.js';
+export { SirenField } from './siren-field.js';
+export { SirenLink } from './siren-link.js';
