+++ src/shipwreck.js
@@ -43,7 +43,7 @@ export class Shipwreck extends EventEmitter {
       if (this.target && this.target.contains(e.target)) {
         e.preventDefault();
         const method = e.target.getAttribute('method');
-        if (method === 'DELETE' && !confirm('You are performing a DELETE. This action is potentially destructive.')) {
+        if (method === 'DELETE' && !confirm('You are performing a DELETE. This action is potentially destructive.')) { // eslint-disable-line
           return;
         }
         this.formSubmit(e.target);
